requirements.txt
altair==4.2.0
anyio==3.6.2
appnope==0.1.3
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
debugpy==1.6.3
decorator==5.1.1
deta==1.1.0
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.4
executing==1.2.0
extra-streamlit-components==0.1.56
fastapi==0.87.0
gitdb==4.0.9
GitPython==3.1.29
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.17.1
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.0
jupyter_client==7.4.7
jupyter_core==5.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
orjson==3.8.1
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.0
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
rfc3986==1.5.0
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.1
starlette==0.21.0
streamlit==1.14.1
streamlit-authenticator==0.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
validators==0.20.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
zipp==3.10.0
